Rename & remove functions directory
